Migrate entry <> namespaced area encoding
