+++ data-model/src/relative_encodings/mod.rs
@@ -3,3 +3,6 @@ pub use path_path::*;
 
 mod entry_entry;
 pub use entry_entry::*;
+
+mod entry_area;
+pub use entry_area::*;
